- Fix example app
