+++ DebugMenu/DebugMenuExample/DebugMenuExampleApp.swift
@@ -11,25 +11,6 @@ import DebugMenu
 @main
 struct DebugMenuExampleApp: App {
 
-    init() {
-        buildDebugMenu()
-    }
-
-    func buildDebugMenu() {
-        let testToggle = DebugToggleAction(title: "Test Toggle", userDefaultsKey: "testKey")
-        let anotherToggle = DebugToggleAction(title: "Another Toggle",
-                                              userDefaultsKey: "secondKey",
-                                              onToggleComplete: { value in  print("Toggled! \(value)")})
-        let testButton = DebugButtonAction(title: "Test Button", action: { print("Button Tapped") })
-        let testSubmenu = DebugSubmenuAction(title: "Test submenu", dataSource: TestDataSource())
-        DebugMenuStore.shared.addActions([
-            testToggle,
-            anotherToggle,
-            testButton,
-            testSubmenu
-        ])
-    }
-
     var body: some Scene {
         WindowGroup {
             ContentView()
